move GConsult to common
