+++ src/climatevision/generator/business2030/dataclasses.py
@@ -2,11 +2,6 @@
 
 from dataclasses import dataclass
 
-from ..inputs import Inputs
-from ..utils import div
-from ..common.g import G
-from ..business2018.b18 import B18
-
 
 @dataclass(kw_only=True)
 class Vars0:
@@ -27,54 +22,6 @@ class Vars0:
     invest_com: float = None  # type: ignore
     invest_pa: float = None  # type: ignore
     invest_pa_com: float = None  # type: ignore
-
-
-@dataclass(kw_only=True)
-class GConsult(G):
-    emplo_existing: float
-    ratio_wage_to_emplo: float
-
-    @classmethod
-    def calc(cls, inputs: Inputs, b18: B18) -> "GConsult":
-        fact = inputs.fact
-        ass = inputs.ass
-        entries = inputs.entries
-
-        invest = (
-            fact("Fact_R_G_energy_consulting_cost_appt_building_ge_3_flats")
-            * b18.p_nonresi.number_of_buildings
-        )
-        invest_pa = invest / entries.m_duration_target
-
-        invest_com = invest
-        invest_pa_com = invest_com / entries.m_duration_target
-
-        cost_wage = invest_pa
-        ratio_wage_to_emplo = fact("Fact_R_G_energy_consulting_cost_personel")
-
-        demand_emplo = div(cost_wage, ratio_wage_to_emplo)
-
-        emplo_existing = (
-            fact("Fact_R_G_energy_consulting_total_personel")
-            * ass("Ass_B_D_energy_consulting_emplo_pct_of_B")
-            * entries.m_population_com_2018
-            / entries.m_population_nat
-        )
-        demand_emplo_new = max(0, demand_emplo - emplo_existing)
-        demand_emplo_com = demand_emplo_new
-
-        return cls(
-            cost_wage=cost_wage,
-            demand_emplo=demand_emplo,
-            demand_emplo_com=demand_emplo_com,
-            demand_emplo_new=demand_emplo_new,
-            invest=invest,
-            invest_com=invest_com,
-            invest_pa=invest_pa,
-            invest_pa_com=invest_pa_com,
-            emplo_existing=emplo_existing,
-            ratio_wage_to_emplo=ratio_wage_to_emplo,
-        )
 
 
 @dataclass(kw_only=True)
